# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Orders API",
        "description": "API for managing travel orders",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate API token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|pPOn8gH3..."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/travel-orders": {
            "get": {
                "tags": [
                    "TravelOrders"
                ],
                "summary": "List all travel orders, optionally filtered by status",
                "operationId": "listTravelOrders",
                "parameters": [
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "Filter by travel order status ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TravelOrder"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TravelOrders"
                ],
                "summary": "Create a new travel order",
                "operationId": "createTravelOrder",
                "requestBody": {
                    "description": "Travel order data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "destination_id",
                                    "departure_date",
                                    "return_date"
                                ],
                                "properties": {
                                    "destination_id": {
                                        "description": "ID of the destination",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "departure_date": {
                                        "description": "Date of departure",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-01"
                                    },
                                    "return_date": {
                                        "description": "Date of return",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Travel order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelOrder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/travel-orders/{id}/status": {
            "patch": {
                "tags": [
                    "TravelOrders"
                ],
                "summary": "Update the status of a travel order",
                "operationId": "updateTravelOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Status data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_id"
                                ],
                                "properties": {
                                    "status_id": {
                                        "description": "2 = approved, 3 = cancelled",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    },
                    "403": {
                        "description": "Forbidden - the user cannot update their own travel order status"
                    },
                    "404": {
                        "description": "Travel order not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/travel-orders/{id}": {
            "get": {
                "tags": [
                    "TravelOrders"
                ],
                "summary": "Get a travel order by its ID",
                "operationId": "getTravelOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel order to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelOrder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Travel order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/travel-orders/{id}/cancel": {
            "patch": {
                "tags": [
                    "TravelOrders"
                ],
                "summary": "Cancel an approved travel order, if it's still within the valid time window",
                "operationId": "cancelTravelOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel order to cancel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully"
                    },
                    "400": {
                        "description": "Order is not approved or is already cancelled"
                    },
                    "403": {
                        "description": "Unable to cancel because return date has passed"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/travel-orders/search": {
            "get": {
                "tags": [
                    "TravelOrders"
                ],
                "summary": "Search travel orders by multiple criteria (status, date range, destination)",
                "operationId": "searchTravelOrders",
                "parameters": [
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "Filter by status ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter by departure_date >= start_date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter by departure_date <= end_date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    },
                    {
                        "name": "destination_id",
                        "in": "query",
                        "description": "Filter by destination ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TravelOrder"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "TravelOrder": {
                "title": "Travel Order",
                "required": [
                    "id",
                    "user_id",
                    "destination_id",
                    "departure_date",
                    "return_date",
                    "status_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the client",
                        "type": "integer",
                        "example": 1
                    },
                    "destination_id": {
                        "description": "ID of the destination",
                        "type": "integer",
                        "example": 3
                    },
                    "departure_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-01"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-10"
                    },
                    "status_id": {
                        "description": "Status ID (1 for requested)",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05T15:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05T15:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "TravelOrders",
            "description": "TravelOrders"
        }
    ]
}